refactor: 로그인 데이터 로컬 저장 로직 분리
