+++ src/hooks/useLogin.ts
@@ -3,16 +3,12 @@ import { convertPath } from 'apis/convertURI';
 import { postsignup, postLogin, SignupRequest } from 'apis/auth';
 import React from 'react';
 import useErrorHandler from 'error/useErrorHandler';
-
-interface UserDataType {
-  isAdmin: boolean;
-}
-
-const defaultLoginState = { isLogin: false, token: '', isAdmin: false };
+import { removeLoginData, saveLoginData } from 'utils/loginDatahandlers';
 
 const useLogin = (redirectPage?: string) => {
   const navigate = useNavigate();
   const { apiErrorHandler } = useErrorHandler();
+
   /* ------------ 로그인 요청 부분 ------------ */
 
   const loginBtnHandler = (): void => {
@@ -24,8 +20,8 @@ const useLogin = (redirectPage?: string) => {
   const signup = (requestBody: SignupRequest): void => {
     postsignup(requestBody)
       .then((response) => {
-        // 로그인, 토큰 저장
-        saveLoginData(response.headers.authorization, response.data);
+        // 회원가입 성공 -> 로그인 처리
+        afterLogin(response.headers.authorization, response.data);
       })
       .catch((error) => {
         apiErrorHandler(error);
@@ -36,7 +32,7 @@ const useLogin = (redirectPage?: string) => {
     postLogin({ code: code })
       .then((response) => {
         // 회원일 경우 로그인
-        saveLoginData(response.headers.authorization, response.data);
+        afterLogin(response.headers.authorization, response.data);
       })
       .catch((error) => {
         // 비회원일 경우
@@ -51,37 +47,22 @@ const useLogin = (redirectPage?: string) => {
 
   /* ------------ 로그인 상태 관리 부분 ------------ */
 
-  // const [loginState, setLoginState] = useAtom(loginAtom);
-
-  const saveLoginData = (token: string, userData: UserDataType) => {
-    const redirect: string = sessionStorage.getItem('beforeLoginURL') || '/';
-    sessionStorage.removeItem('beforeLoginURL');
-
-    const loginData = {
-      token: token,
-      isLogin: true,
-      isAdmin: userData.isAdmin,
-    };
-
-    // setLoginState(loginData);
-    sessionStorage.setItem('login', JSON.stringify(loginData));
-    navigate(convertPath(redirect));
+  const afterLogin = (
+    token: string,
+    userData: {
+      isAdmin: boolean;
+    },
+  ) => {
+    saveLoginData(token, userData);
+    navigate(convertPath(sessionStorage.getItem('beforeLoginURL') || '/'));
   };
 
   const logout = () => {
-    sessionStorage.removeItem('login');
-    // setLoginState(defaultLoginState);
+    removeLoginData();
     navigate(convertPath('/'));
   };
 
-  const getLoginState = () => {
-    const stringData = sessionStorage.getItem('login');
-    if (stringData === null) return defaultLoginState;
-
-    return JSON.parse(stringData);
-  };
-
-  return { logout, signup, login, loginBtnHandler, getLoginState };
+  return { logout, signup, login, loginBtnHandler };
 };
 
 export default useLogin;
